+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# Unreleased (1.10.4)
 
 # 1.10.3