+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * Removed SymStackConst support
 * Added variant (union) support for AuxData
+* Removed address and size from modules
+* Modified Module::findSections(string) to return a range instead of iterator
+* Added IR::findSection(string) to C++ API
 
 # 1.10.3