+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased (1.10.4)
+# 1.10.4
 
 * Removed SymStackConst support
 * Added variant (union) support for AuxData